+++ src/components/generics/index.js
@@ -3,4 +3,5 @@ export * from './ITHeader/ITHeader'
 export * from './ITPanel/ITPanel'
 export * from './ITTags/ITTags'
 export * from './ITActions/ITActions'
-export * from './ITModal'
+export * from './ITModal/ITModal'
+export * from './ITTable/ITTable'
